+++ integration_test/theme_picker_test.dart
@@ -31,7 +31,7 @@ void main() {
         'builtin.theme-picker',
         'builtin.default-layout',
       ],
-      daemonClientFactory: (log, events) => FakeDaemonClient(log: log, events: events),
+      daemonClientFactory: (log, events, _) => FakeDaemonClient(log: log, events: events),
       autoStartDaemonClient: false,
     );
     services.extensions